Python/Pyrogram/JokesBot/requirements.txt
requests
pyrogram
python-dotenv
